implemented line search for the new newton method
it is disabled by default, though, since it seems that the algorithm
performs worse both in terms of the number of iterations as well as in
terms of CPU time with line search than without it. It seems that the
squared two norm of the residual is reduced nicely, but maybe
something else is wrong. Weird...