get rid of dumux/boxmodels/tags.hh
since property tags can now be defined multiple times it is possible
to include multiple models which all define the same property
tag. thus, by moved the definition of all property tags to the
respective models, the compiler now complains if one tries to specify
a property which makes no sense for the respective model.