fix a capital bug in the newton method
for iterative linear solvers we have to reset the delta vector to zero
every time, because the ISTL solvers use the reduction of the residual
as the convergence criterion. if the newton method is converged the
delta vector is 0. Now if the delta is the solution of the last
iteration the initial defect can be extremely large (e.g. 10^10) so a
the relative reduction which needs to be achieved by the linear solver
would have to be absurdly high!