box FV element geometry: make it independend of dune-pdelab
This also has the advantage, that hybrid grids should work again from
the FVElementGeometry's POV. basically the only dependence on PDELab
which is left is the linear solver. This dependcy is easy to get rid
of for the sequential case, but much harder for parallel computations
(at least for clueless persons such as me).