box models: move handling of the extrusion factor to the base classes
the extrusion factor describes how much a 2D or 1D is extruded to
represent the physical 3D domain. by default, it is 1 m to the back
for 2D models and 1m^2 for 1D models. Usually this factor does not
matter too much, but when the absolute amount of stuff is relevant, it
does. Examples are radial symmetrical domains approximated in 2D or a
coupled 1D-3D model.

this makes all box models respect the extrusion factor. (before, it
was only the Mp-Nc model.)