Parameters: (re-) introduce runtime parameters
runtime parameters are just like normal parameters, except that they
don't have a property as a fallback. Thus, they are mandatory. Any
parameter can now be either a normal parameter, or a runtime
parameter. If a parameter is queried as a runtime parameter as well as
a normal one, an exception will be thrown in debug mode.