property system: improve introspection
now it is possible to print where the point of definition and the
inheritance structure for all properties of a given type tag. for
properties defined using the convenience macros SET_*_PROP it is even
possible to print their value. the generic startup routines plus
test_2p have been adapted to print everything. todo: introduce a
parameter which allows to turn of the output at runtime.