boxmodels: allow using two-point gradients
if twopoint gradients are enabled, the FE gradients are not calculated
anymore. since the Mp-Nc model is the only one which really implements
two-point gradients, all other models will fail if this property is
set to true.

TODO: introduce a generic "gradient model". also a generic "velocity
model" would be nice which could be Darcy's law or the Forchheimer
equation...