make the peng-robinson EOS faster by tabulating the critical points
For cubic EOS, the critical values only depends on the attractive
paramter "a" and the co-volume of the fluid "b" which makes them an
easy prey for tabulation. also added: a class which does bi-linear
interpolation on an arbitrary function.

TODO: extend Dumux::Tabulated2DFunction so that upper and lower limit
      curves f(x) and g(x) can be specified and the sampling points do
      not need to be equidistant anymore. If this is done, use it for
      Dumux::TabulatedMaterial2.

MORE TODO: In Dumux::Tabulated2DFunction, do not use linear
           interpolation, but take the partial derivatives at the
           sampling points into account. then, interpolate the partial
           derivatives linearly between the sample points. I have not
           spend too much time thinking about the, though...