revert r6411 with some modifications
the reason why the boundary indices where a global array is because it
is possible that an element's vertex is on the boundary, but the
element does not have any boundary intersections. only checking for
boundary intersections does not work for dirichlet conditions on such
grids. (for simplex grids that case is actually quite common.)