remove Dumux::inDebugger()
running ptrace(PTRACE_TRACEME, ...) on a process always causes the
Unix signals to be intercepted, i.e. after pressing Ctrl+C the process
did not terminate itself anymore, but it was merely stopped. since
this is really undesireable, expecially for simulations which use
large quantities of memory, and I couldn't come up with a way to
detect whether a debugger is attached that does not have this
drawback, I figured that it is better to always catch the exceptions
thrown by the simulation. To make debugging possible, one can either
set a breakpoint in __cxa_throw() or use GDB catchpoints:

http://www.delorie.com/gnu/docs/gdb/gdb_31.html