2p2c box model: cleanup
the primary variable switch is now implemented independent of the
number of phases, there is now only one formulation (which is
determined implicitly), the PV-switching code has moved to the primary
variables and is quite a bit shorter (although not simpler),
everything has been converted to moles (i.e. the model now uses mole
fractions instead of mass fractions, and the fluxes/storage terms are
specified in moles).

the only 2p2c specific code which remains in the model is the volume
variables. Replacing it with something generic is on the radar but not
too high on the priority list.