rework printed messages
Notable changes:
- we now support a code name
- eWoms is not an airline anymore
- in the box models, assemble/solve/update timings for the whole
  simulation are printed in addition to the individual time steps.

I've also noticed that the split of the Newton method into the Newton
algorithm and a Newton controller should probably be revised in the
medium term, as it makes things harder than strictly necessary. (The
original idea was to be able to provide completely different newton
controllers for different models without suffering the memory overhead
of a common base class. As it turned out, all current controllers just
inherit from the default controller and override a few methods. The
same can be done if everything is collapsed into one class.)