fluid states: modularize them
now, each group of quantities can be stored in a number of ways and
the fluid states can combine these modules arbitrarily. The following
groups of quantities have been defined (so far):

- temperature
- pressure
- composition
- fugacity/chemical potentitial
- density
- viscosity
- enthalpy

For all of these groups, at least two modules are provided: a module
which stores all quantities explicitly, and a null which just throws
Dune::InvalidStateException if the quantity is accessed. In addition
the following modules are provided:

- immiscibility for composition and fugacity
- equilibrium for temperature