stokes: cleanups, impose boundary conditions weakly
the "stokes" model up to this point was _not_ discretizing the stokes
equations, because it consideres storage of momentum. I guess it was
rather "Navier-Stokes" with a few terms missing.

Also, the boundaries used to be "stabilized" in a very weird way. no
more: Boundaries are handled just like in any other model now...

Finally volume variables may now contain gradients in all box models.