box models: make them work in 3D, remove numFAP from BoxFVElementGeometry
the centers of the sub controlvolumes where only calculated for 2D
grids, now it's also done for 1D and 3D grids. For 3D grids pyramids
and prims are still missing. Also fix a bug in the
BoxMultiPhaseFluxVariables which prevented it from working in 3D.

TODO: The BoxFVElementGeometry is in serious need of a cleanup!