the default of the beta scaling parameters did not work. Now fixed.
What are the beta scaling parameters?

They allow a scaling of two-phase capillary pressures according to
the surface tensions. Typically, the Gas-Water system is the reference
and the betaGW parameter is 1. Surface tensions between the other
fluid phases are a little different, and this can be taken into
account phenomenologically by the beta scaling parameters. An example
is given in dumux-lecture/mm/columnxylene exercise.

Cherry picked from SVN trunk (revision 8424)