grid creators: properly clean up before exit
if the grid is not explicitly deleted, then it will be done implicitly
at exit. the problem is that the constructor of Dune::MPIHelper is
also called implicitly at exit. For destroying the grid the MPIHelper
must still exist, but the order in which the destructors of static
objects are called is undefined by the C++ standard...