immiscible box model: fix stupid bug in globalPhaseStorage()
the size of the finite volumes was not accounted for. if the source
term also did not consider this, the results of globalPhaseStorage()
looked perfectly fine even if they where orders of magnitude of. (for
large cells, this resulted in very large pressures, which burned me
for the Ketzin problem.)