algebraic overlap: introduce "native indices"
native indices are the old local indices (i.e. indices that are local
to a given process) and local indices are now defined as indices that
are local to the process and do not contain any blacklisted
indices. blacklisted indices stem from grid overlap or ghost entities
and are not required for the linear solver. If they are not removed,
the AMG gets into trouble for some reason, though.