parameter system: major overhaul
"runtime" parameters are gone now, as are groups. both where flawed
concepts: parameters can be overwritten at runtime by definition, and
a parameter that has no compile-time fallback just degrades
performance and is a hassle to deal with because it has to be
specified in a parameter file. The other killed feature, parameter
groups, just made it harder to figure out what the correct command
line argument was to change the parameter (was it
--linear.solver-tolerance, --linear-solver.tolerance or
--linear-solver-tolerance?) and did not provide much value otherwise.

as a consequence all parameters now have compile-time defaults which
allows us to remove the auto-loading of a parameter file at
startup. (These files became quite obnoxious in eWoms when problems
declared their independence from models.)

Finally all parameters must now be registered before they can be used
which allows to properly include them in the usage message.