linear solvers: cleanups, introduce an absolute tolerance
the absolute tolerance allows the linear solver to converge, even if
the actual criterion (e.g. reduction of the residual) is not
fulfilled. This is useful for for small defects where the linear
solver runs into precision problems. Note that, since "small defect"
is a problem specific property, this criterion is disabled by default.