decoupled: provide default initial() method again
Since commit r10693, all problems that use a model based on FVPressure
had to provide an initial(AtPos) function. Default implementations are
available in onemodelproblem.hh and impetproblem.hh, but threw an
exception. Since many test problems have lived for several years with
an initial guess of 0, the default behavior has been changed to not
throw an exception and to set the values to 0.

Dumux-Svn-Revison: 10737
Ported-By: Andreas Lauser <andreas.lauser@iws.uni-stuttgart.de>