make all epsilon values settable in the PVS and NCP models
also make the base epsilon used for finite differences settable. So
far, these values can only be modified at compile time as they are
very low level (it is strongly discuraged to mess with them if it can
be somehow avoided) and they are also very preformance critical,
i.e. the compiler may loose a few optimization opportunities if these
values are only available at run time. the motivation for this is to
make it possible to run SPE-5 which for some reason seems to be very
sensitive to this...