avoid deprecation warnings in dune 2.3
these were caused by the rename from "GenericReferenceElement*" to
"ReferenceElement*". Note that we actually test for the version of the
dune-common module instead of the version of dune-geometry. The reason
for that is that I don't want to change the Findewoms.cmake CMake
script because that would need to be distributed to all OPM modules,
which is a lot of work...