fix some initialization order issues
basically the grid view was not available until too
late. Unfortunately, this requires some API changes and this
initialization is still a beast because whether one gets defined
results now depends on the order of initialization of the problem.
(this is not too much of an issue though, as all base problems for
models derive from the same class.)