fix compilation with old versions of GCC
these require a copy constructor for objects to be usable with
std::vector even if the constructor is not actually called. Since this
constructor was explicitly deleted, so the build failed.

This patch is very specific for GCC, but since the issue does not
appear with more modern versions of it and also not with clang, I do
not want to bother to write a build system test for this. Also, we
provide a copy constructor for GCCs up to 4.6, but I could not test it
on GCC 4.5 and 4.6, so the fix may not be necessary for them. (but
this does not really matter since all development should be based on
more modern compilers anyway...)