add a timer class better than Dune::Timer
"better" because it differentiates the between CPU and real time and
it can also properly consider parallel simulations. Also it seems to
have a higher resolution as I don't get negative "overhead" with this
anymore. The only downside which I can see is that the code is
currently Linux specific (or POSIX, I don't know).