reverse the assembly to stand on its feet instead of its the head
i.e. the local Jacobian now calculates the derivatives of the residual
functions with regard to the primary variables of the "primary DOFs"
instead of the derivative of the residuals of the "primary DOFs" with
regard to all primary variables of all degrees of freedom in a stencil.

this has the advantage that the residual has to be evaluated only once
per primary DOF and the "change of residual" for the other DOFs come
for free because this is just the fluxes which go through a face. For
discretizations where the number of primary degrees of freedom in a
stencil is significantly smaller than the total number of DOFs, this
leads to a significant speedup of assembly time. (most importantly,
this applies to the element centered finite volume discretization if
using two-point gradient approximation.)