improve gravity correction
instead of adjusting the pressure at the center of one DOF to the
height of the center of other DOF, we now correct both centers to be
at the same height as the interface between the DOFs. this has the
advantage that it is symmetric in the sense that it does not matter
anymore which DOF represents the inside FV and which one the outside
one. (if the fluids are compressible, this matters.)

this makes the results produced by ebos to fit within 2% of those of
eclipse for SPE1...