ECL well manager: hack around recently introduced opm-parser ideosyncrasy
catching an exception for this seems like a pretty bad hack to me, but
there seems to be no other way to detect that a deck did not specify a
completion radius. (well, one could look at the raw COMPDAT keyword,
but that would defeat all benefits of using opm-parser's schedule
objects.)

this makes ebos work with SPE9 for the current master version of
opm-parser again.