fix and improve the partial relinearization code
this is a very simple idea, but the implementation turns out to be
very complex: the right-hand side is irrelevant for partial reassembly
because the update of the solution can be large even if the right hand
side is zero. This means that the only relevant quantity for deciding
what to update is the deflection of the primary variables between
iterations.

The reason why this worked before, is that the right-hand side was
recomputed unconditionally and the left hand side often did not change
much (or in some instances at all). With this patch, the RHS is not
updated anymore if it is not necessary.