partial relinearization: always update the solution
and always fully recompute the right hand side but only update the
left hand side if the deflection from the linearization point is too
large.

not updating the solution and the RHS lead to very bad convergence
rates. now they seem to be basically the same as without partial
relinarization...