call the 'stty' command to reset the terminal to a useful state
before this, we tried to do this ourselfs, but it seems like we missed
something so the code was unrelieable. Let's see if this is an
improvment, but at least we do not need the termios.h header anymore,
and all of the eWoms code is GPL 2+ again. (the 'stty' command from
which the previous code was ripped off is GPL 3+... (well, "all code"
means "all code except for the solvers.hh file" which is licensed
using the GPL2 with the Dune template exception, so it effectively is
LGPL2...))