add a Ewoms::declval() template function
this function returns an object given an arbitrary type (which does
not require to exhibit a default constructor) and is hopefully 100%
equivalent to std::declval(). This function is useful in conjunction
with the c++ 2011 decltype() keyword to determine the type of the
return value of non-static member functions

The only reason to add this function to Ewoms are old compilers like
GCC 4.4 which do not provide it in their standard library.