improve the forwarding of types in the quantity callbacks
in particular, we now don't convert references to values anymore. So
far, this did not matter because the only performance-critical type
which was returned by these classes where Scalars, but now these
objects can be function evaluations and returning these by value can
be be considerably slower than returning them by reference.

also, whether to use two-point gradients for the VCFV discretization
or not must be specified at compile time (instead of being a runtime
parameter). The reasons are that this is an option which is very deep
in the guts of the simulator (so it is a probably not a too good idea
to make it so easily tweakable) and the also that two-point gradient
calculator can potentially be more performant if it can return a
reference instead of a value.