remove the concept of an "evaluation point"
this concept is specific to finite differences and does not improve
things significantly (if at all). In addition to this, I'm not even
sure if it was implemented correctly. Besides the additional code
complexity and the associated maintainance burden, keeping track of
"evaluation points" requires considerable effort, so that the runtime
performance may even slightly improve by removing this.