adapt to yet another ERT API change
the unit conversion of ERT is a bit weired: internally, it always
assumes that lengths are metric and I haven't seen any way to change
it. Anyway, if we alway tell ERT the units are metric and deal with
the conversion ourselfs, everything is as before. (at least until this
gets changed in ERT again...)