remove the ability to select a level grid view
this code was buggy and thus did not work. since Dune's definition of
a grid level is also pretty counter-intuitive (IIRC, a level
represents the delta of the entities compared to the next-lower level,
but I could very well be wrong), let's remove this functionality until
it is needed by something.