OverlappingBCRSMatrix: simplify the index exchange communication code
since the calculation of the row and (row, column) indices are now
decoupled from the MPI communication and the job of discarding entries
for columns which are not in the overlap is now moved to the receiver,
the code hopefully now is quite a bit easier to understand and more
robust.

this should fix the computation of the distributed Jacobian matrix for
the element-centered finite volume discretization on grids which only
provide ghost elements (i.e., ALUGrid and UG).