make the linear solver work for arbitrary overlaps
hopefully: the problem was that in some instances, some degrees of
freedom actually where in the front of the overlap but their border
distance as computed by the foreign overlap is smaller than the
overlap size. the operator of the solver needs to synchronize such
DOFs from their master ranks, but it did not. the way out taken by
this patch is to always do a full synchronization after each
application of a linear operator. this leads to more (for large
overlap sizes, considerably more) network traffic, but we can be sure
that the resulting vectors are correct. (also, the "front
synchronization" stuff can be removed from the OverlappingBlockVector
class which significantly reduces its complexity. *yay*)