implement caching of the storage term
once again, this is a bit hacky because it assumes that an Euler time
discretization is used and that the solution for the time indices 0
and 1 are identical at the beginning of a time step. Nevertheless, it
saves us from updating the intensive variables for solutions other
than the one of the current iteration.