make caching the storage term work in all cases if the intensive quantities cache is enabled
it did not work properly if the time integration failed. besides that
bugfix, make it faster by taking advantage of the fact that the
intensive quantities for the first iteration of a time step do not
need to be calculated because the solution for this iteration is the
same as the solution for the last time step. (this only affects
simulators like ebos which need go over the whole solution as a
post-process step and thereby update the intensive quantities. In the
case of ebos, the well rates are calculated there.)