ebos: notify the user about parallelism and Dune::CpGrid in a better way
now, we abort at run time if the user tries to use with Dune::CpGrid
in parallel simulations. The advantages are that during build there
will be no warning anymore (before the warning was printed at compile
time regardless of which grid was actually chosen), and that such
simulations cannot be started with more than a single process (before
they started just fine, but they did not work properly).

This code should be removed as soon as "Dune::CpGrid" gets its act
together and fixes the bug in the loadBalance() method.