black-oil local residual: fix problems with the reference densities
now, the reference density is always taken from PVT region of the
upstream element of the respective phase, not from the PVT region of
the upstream element of gas. this issue one only bit occassionally and
only for decks which enable capillary pressure and use more than a
single PVT region.