flash model: don't screw up with the thermodynamic hints
this prevents the cat to chase its own tail, i.e., the temperature
must always be taken from the primary variables. as the comment above
this line clearly indicates this was not the first time I stumbled
over this problem, but when I was did the "Local-AD enablement" I was
in bulk mode and in all other places a const reference to the
Evaluation could replace a Scalar object, so this slipped through the
cracks.