allow lockless multi-threaded linearizations
this means that it is possible for the discretization to specify that
aquiring a lock is not required when assembling the global residual
and its Jacobian matrix. this is the case if an entry of the grid-wide
vector/matrix is only written a single time. (most relevantly here,
this is the case for the element-centered finite volume
discretization.)

setting this property to "true" unconditionally is safe from the
correctness point of view, but it may escrew a bit of performance if
the linearization is done in a highly multi-threaded mode. (Since this
requires to enable OpenMP at build time and to explicitly set the
ThreadsPerProcess parameter to a large value at runtime (>= 16?), this
is not something which should affect many people, but its nice to have
anyway.)