black-oil: specify the phase presence explicitly
it is now based on the interpretation of the primary variables instead
on the value of the phase saturation. In principle this should not
make a difference but since the BlackOilFluidState distinguishes
between saturated and undersaturated conditions it does if a variable
has just been switched. the main purpose of this patch is to make
everything identical to what the flow simulator does...