black-oil: change the primary variables back to { Sw, pg, Rv } for the gas-water case
this fixes a inconsistencies when determining if the oil phase appears
and thus should reduce oscillations of the newton method. On the
flipside, the method which adapts the primary variables is now more
elaborate because it needs to calculate the capillary pressures on
switches and it is conceptually less pure because it needs to ask the
problem directly via the global index of the DOF (because a full
context object is not available/probitively expensive at this place).