FvBaseLinearizer: provide non-const versions of residual() and matrix()
these make it a bit simpler to implement the well model of flow_ebos
directly on top of the linearization which is produced by ebos.

I'm not completely happy with modifying data structures which the
model "owns" this way, but I'm not dogmatic about it. E.g., modifying
the matrix externally makes it next to impossible to implement tricks
that reuse the old linerization (like "partial relinarization"), but
currently no such things are present in the ewoms code base, so I
guess that this patch currently will not cause trouble.

(I just noted that matrix() is a very bad name for the method. it
should be called 'jacobian()' but I'll leave that for another day.)