overhaul the implementation of constraint degrees of freedom
instead of trying to apply constraints implicitly during the
linarization, we now do an explicit pass over the grid before the
"real" linearization.

as an optimization, this is only done for the first iteration of each
time step, i.e., constraints can be time dependent, but not solution
dependent. (i.e. degrees of freedom with "partial" constraints are not
possible.)