NewtonMethod: make it possible to abort after the linearization stage of an iteration
for some convergence criteria (i.e. the ones which only look at the
residual but not at the delta of the solution updates), this allows to
not having to solve the linearized system of equations. For ebos on
SPE9_CP, performance improves approximately 7% on my machine. (new:
15.50 seconds; old: 16.60 seconds.)

Note that this change implied some refactoring of the linear solver
API in order to keep it working for the MPI-parallel case.