Timer: switch to std::chrono
this makes a few annoyances due to buggy target systems go away, but
on the flipside, the CPU time measurements get a few orders of
magnitude less precise on Linux systems. (the measurements of real
time should stay unaffected, though.)

medium-term this class should be replaced by Dune::Timer in my
opinion, but that would be a more invasive endeavor.