make it possible to specify a separate floating-point type of the linear solver
The floating point type which ought to be used internally by the
iterative linear solvers can now be specified via the
"LinearSolverScalar" property. It defaults to "Scalar", so this patch
does not change the default behaviour.

The most common case of this is probably to use single precision
floating point scalars for the linear solver and to do the
linearization using double precision. For the lens_immiscible_ecfv
test problem with two grid refinements, the performance improvements
seem to be considerable: On my machine the linear solver speedup is
almost 50%. On the flip-side, there are precission problems, so the
tolerance of the Newton-Raphson solver must be relaxed when using
single-precision for the linear solver.

The exact command which I used for that measurement is:

./bin/lens_immiscible_ecfv --newton-raw-tolerance=1e-6 --grid-global-refinements=2

In the future, making the linear solver automatically switch between
`float` and `double` might solve the precision problems, but I need to
think a bit about how to implement this.