WeightedResidReductionCriterion: disable the fixpoint criterion
the idea of this was to stop the linear solver early if the solution
does not change anymore. unfortunately this cannot be done so easily
because the implementation assumed the exponent of the floating point
number to be >= 1 which leads to a way too large tolerance if the
scalars are single-precision.

interestingly, this patch also seems to have a (smaller) positive
effect if double precision scalars are used.