set the absolute convergence criterion to the a tenth of the Newton tolerance
there's not much point in the linear solver being much more precise
than the Newton-Raphson method: in eWoms the absolute convergence
critererion of the linear solver is the same as the convergence
criterion of the Newton-Raphson method. The only difference is that
due to the non-linearity of the equations, in general the residual of
the linearization will be different from the non-linear one when
deflecting the solution from the linearization point. (i.e., if a
system of linear equations is solved by the Newton-Raphson method,
both kinds of residuals would be exactly the same.)