parallel itererative solvers: increase the maximum number of iterations to 1000
this makes the linear solver fail less quickly. The reason is that
some of the tests converge better in this case and that compared to
AMG solvers, iterations are pretty cheap. For the lens_immiscible_ecfv
test with two grid refinements this significantly reduced the number
of failed time steps.