refactor the linear solver backends and the convergence criteria
this gets rid of quite a bit of copy-and-pasted code, replaces the
"equation weight vector" for the CombinedCriterion class by rescaling
the linear system of equations to be solved, improves the API for
convergence criteria with an eye on performance and gives the
preconditioner- and solver wrappers their own files.

note that with this the prestine linear solvers from dune-istl now
also respect the weight of equations (due to the rescaling the linear
systems before solving them). Interestingly, the prestine dune-istl
solvers do not seem to profit from this performance-wise while the
home-brewn `BiCGStabSolver` seems to get 7 to 8% faster. (I.e., its
performance advantage compared to `Dune::BiCGSTABSolver` gets close to
50% for my testcase.)