fix a regression for the parallel linear solvers
the main problem that the stagnation detection used the maximum
instead of the minimum over the processes (i.e., the convergence
criterion thought the solver stagnated if there was at least _one_
process where the solution did not change instead of _all_ processes).

also, there was the rescaling code did not translate domestic indices
into native ones, but this did not cause the regression.