simulator: add a way to force the time step sizes externally
the forced time steps are read from a file which is supposed to
feature one time step size in seconds per line. this kludge is
intended to ease debugging/comparisons, so it does not even attempt to
cover any corner cases (like e.g. non-convergent time steps). given
that this feature is turned of by default, it should not hurt anyone
to have it?!

v2: read the whole file into a std::vector at init. thanks to
    [at]atgeirr for the suggestion!