Make it possible to formulate the blackoil eq in terms of surfacevolumes
One tag that formulates the equations in terms of surface volumes
instead of mass.
NEW_PROP_TAG(BlackoilConserveSurfaceVolume);

and one to specify whether to use volumetric residuals or not
NEW_PROP_TAG(UseVolumetricResidual);

If both are enabled the blackoil equations will be formulated in terms
of surface volumes instead of mass/volume. (default)