NCP primary variables: make assignNaive() more robust
the old version of this method required the fugacities of the fluid
state to be correctly defined. since fugacities are somewhat exotic
quantities, and thus often not defined for the initial condition, we
now compute them using the composition of the reference phase. This
makes the code somewhat slower and slightly less generic -- before it
was possible to only define the component fugacities in the reference
phase but not its composition -- the new code should be vastly more
bullet-proof and less surprising.