fix the build on GCC 4.8
it seems like GCC 4.8 has some issues with the SFINAE rule (i.e.,
std::enable_if). Anyway, checking these conditions at runtime using a
normal if() statement should not have a negative impact on performance
because the compiler should notice that the condition is a constant
expression which is specified at compile time and eliminate the dead
branches. (in fact, its probably even beneficial to do it this way
because all the code is seen by the compiler.)