Only apply the "primary variable switching hysteresis" to the saturated values of the dissolved components
this makes it unsymmetrical if the DRSDT or DRVDT keywords are
present, but when using PV switching hysteresis, SPE-1 case 1 does not
work and when not using it for the saturated Rs and Rv values,
realization 5 of "model 2" exhibits significant performance regressions.

See https://github.com/OPM/ewoms/pull/271 for the full discussion.