write the VTK output asynchronously by default for sequential simulations
`Dune::VTKWriter` seems to be on the rather slow side, so threaded
output writing leads to a sizeable speedup for the VTK format. For
situations where an extra thread is not desireable, asynchronous VTK
output can be disabled using the `--enable-async-vtk-output=false`
command-line parameter.

The main complication of this is that `Dune::VTKWriter` cannot be used
in conjunction with MPI-parallel simulations because there are
unnecessary calls to `MPI_barrier()` in `Dune::VTKWriter::pwrite()`
which lead to all kinds of fun since multiple threads are doing
uncoordinated MPI calls in that case. (This is because the main thread
needs to call MPI functions during e.g. the linear solver.) That said,
for the MPI-parallel simulations which I've tested, there was no
speedup either, so asynchronous output is simply disabled for the MPI
case.