possibly fix a race condition in the tasklet mechanism
if `barrierCondition_.notify_all()` happened while another thread was
preparing to call `barrierCondition_.wait()`, the preparing thread
would miss the notification and `barrierCondition_.wait()` would lock
up. after this patch, this should not happen anymore because the whole
wait() method of the barrier tasklet is operated with the barrier
mutex locked. This has the added benefit of making the code a bit
simpler.

this race only occured if condition variables do not check the
predicate before waiting. My references not very explicit about this.