Make the "BlackOilEnergyScalingFactor" a "static const" variable
some versions of GCC have problems with `constexpr` if optimizations
are disabled (my GCC 5.4 produces linker error if trying use such
constexpr variables directly). using `static const` essentially boils
down to the same end, but it requires to define the variables outside
of the class, so some low-level macros have been added to the property
system to generate the boiler plate code for this.