fix up the more uncommon convergence criteria for the default linear solver
some header guard macros were not adapted after copying the respective
source files and ´WeightedResidualReductionCriterion::update()` was
not adapted after the signature of the method changed in the base
class. the practical implications of this were limited, because the
affected convergence criteria are not used by default. (i.e.,
currently they are purely used for debugging purposes.)