parameter system: allow parameters to be specified only once per location
i.e., defining the same parameter multiple times on the command line
or the parameter file will cause the run to be aborted. it is still
possible to specify a parameter on both, the command line and the
parameter file, though. (in this case, the value of the command line
will take preceedence.)