black-oil: minor fixes to the update method
this primarily updates the comments in
`BlackOilNewtonMethod::update()` and it prevents using
`EWOMS_GET_PARAM()` in a fast path: EWOMS_GET_PARAM() needs to
traverse a `std::map` of strings every time, so it should not be used
in highly performance critical code. Given that, it is surprisingly
fast, so the performance improvement of this is probably quite small.