Fix eclipse output for reordered CpGrid and add command line parameter.
The eclipse output expects the arrays in CellData to be ordered correctly, that
is local indices should be such that cartesian index i is running fastest, then
j and then k. In case the CpGrid is reordered that does not hold. Therefore we reorder
the array just before calling the eclipse IO methods.

In addition we added a command line parameter --enable-reorder-k-fastest to switch on
reordering. Default is false.