Handle exceptions throw in OpenMP threads
Exceptions cannot escape from OpenMP threads to be handled at the outer
level in the master node. If an exception occurs and is not handled
within the parallel block, terminate() will be called!

This patch catches all exceptions that occurs in threads, tuck them away
and let the parallel block end normally. Afterwards one of the
exceptions are selected at random and then rethrown in single thread mode.