blackoil model: fix the equation weights
The main change is that we now care quite a bit less about water than
about the hydrocarbons. Also, the equation weight was inverted when
conserving fluid surface volume instead of mass. Note that the
tolerances for oil, gas and water are now always expressed in terms of
"invalid kg per second", i.e., the tolerances for the Newton-Raphson
solver probably should be changed for problems that used the black-oil
model formulated in terms of surface volumes.

this is only relevant for the eWoms Newton-Raphson solver, i.e. `flow`
is completely unaffected at the moment.